requirements.txt

asgiref==3.4.1
backcall==0.2.0
beautifulsoup4==4.10.0
bootstrap4==0.1.0
decorator==5.1.0
Django==4.0
django-bootstrap4==21.1
django-crispy-forms==1.13.0
django-environ==0.8.1
django-filter==21.1
django-rest-framework==0.1.0
djangorestframework==3.13.1
ipdb==0.13.9
ipython==7.30.1
jedi==0.18.1
matplotlib-inline==0.1.3
mysqlclient==2.1.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
prompt-toolkit==3.0.24
ptyprocess==0.7.0
Pygments==2.10.0
pytz==2021.3
soupsieve==2.3.1
sqlparse==0.4.2
toml==0.10.2
traitlets==5.1.1
wcwidth==0.2.5
